StringProcessing: Reorganize modules
This commit changes the StringProcessing module in following ways:
- Move StringProcessing functions into the `Core` module inside a new
  StringProcessing directory.
- Move the filters from StringProcessing (`limit`, `trim_empty` and
  `trim_empty_matches`) to `StringProcessing.Filters`.
